// SETPPBO/ClientApp/app/main/administrasi-banding-gugatan/models/administrasi-banding-gugatan/administrasi-banding-gugatan.ts
export class AdministrasiBandingGugatan {
        PermohonanId:number;
        Nama: string;    
        NoTandaTerimaSubSt: string;
        NoSuratPermintaanSubSt: string;
        NoSubSt: string;
        NoSuratPermintaanBantahan: string;
        NoSuratBantahan: string;
        NoSuratPermintaanSalinan: string;
    }
    